% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfilter.R
\name{daily_routine}
\alias{daily_routine}
\title{Observe daily trading records of specific broker}
\usage{
daily_routine(url, folder = "fil_daily")
}
\arguments{
\item{url}{url of google drive folder}

\item{folder}{folder code, 'daily' for Frank;
'fil_daily' for filtered stock.}
}
\description{
Observe daily trading records of specific broker
}
\details{
1. Stop crawling records and rename google spread sheets which were marked in "close" column.

2. Renew existed files.

3. Get trading records for new selected stock.
}
\examples{
url='https://docs.google.com/spreadsheets/d/1_n9Ba3rzswv0L8gtETRXMFv-njcTTLKFLt3WozQCzJQ'
daily_routine(url,'fil_daily')

}
\seealso{
\link{ob_recent}
}
