% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfilter.R
\name{ob_recent}
\alias{ob_recent}
\title{Observe recent trend of chip}
\usage{
ob_recent(url, folder = "fil_recent")
}
\arguments{
\item{url}{url of google drive folder}

\item{folder}{folder code, 'recent' for Frank;
'fil_recent' for filtered stock.}
}
\description{
Get trading records from wangoo and put them into google spread sheet.
Move files info specific folder in google drive.
}
\details{

}
\examples{
url='https://docs.google.com/spreadsheets/d/1_n9Ba3rzswv0L8gtETRXMFv-njcTTLKFLt3WozQCzJQ'
ob_recent(url,'fil_recent')

}
\seealso{
\link{daily_routine}
}
